Updates default URLs per recommendation from Google
Google has recommended updating authentication endpoints
to those found in this commit.

Reference:
- https://github.com/googleapis/google-cloud-common/issues/260
- https://github.com/googleapis/google-auth-library-ruby/issues/155

Some evidence shows old URLs are causing intermittent issues; for
example, an elusive `invalid_grant` message was showing up for a
project I worked on recently, and disappeared with this change.
Possibly an issue with timeouts and/or deprioritization of these
older URLs, or something else that isn't immediately apparent.